[#175973060] refactoring
