+++ ts/features/bonus/bpd/saga/networking/prefetctBpdDetails.ts
@@ -9,7 +9,7 @@ import { BackendBpdClient } from "../../api/backendBpdClient";
 import { isReady } from "../../model/RemoteValue";
 import { bpdLoadActivationStatus } from "../../store/actions/details";
 import { bpdPeriodsAmountLoad } from "../../store/actions/periods";
-import { BpdPeriodAmount } from "../../store/reducers/details/combiner";
+import { BpdPeriodAmount } from "../../store/reducers/details/periods";
 import { bpdLoadAmountSaga } from "./amount";
 import { bpdLoadPeriodsSaga } from "./periods";
